% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_quickglance.R
\name{mod_quickglance_ui}
\alias{mod_quickglance_ui}
\alias{mod_quickglance_server}
\title{quickglance UI Function}
\usage{
mod_quickglance_ui(id)

mod_quickglance_server(input, output, session)
}
\arguments{
\item{id}{shiny id}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}
}
\description{
quickglance UI Function
}
\keyword{internal}
